% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suite.R
\name{agent_suite}
\alias{agent_suite}
\title{Collect cases into a suite}
\usage{
agent_suite(...)
}
\arguments{
\item{...}{Objects of class `r4agent_case`, or lists of them.}
}
\value{
An object of class `r4agent_suite`.
}
\description{
Collect cases into a suite
}
