% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suite.R
\name{as.data.frame.r4agent_suite}
\alias{as.data.frame.r4agent_suite}
\title{Turn a suite into a data frame}
\usage{
\method{as.data.frame}{r4agent_suite}(x, ...)
}
\arguments{
\item{x}{An `r4agent_suite`.}

\item{...}{Ignored.}
}
\value{
A data frame, one row per case.
}
\description{
Turn a suite into a data frame
}
